Add filtering by top 10
I'm happy!

The server computes and attaches custom data attributes to each stream HTML
element about whether or not they fall within a certain filter
(e.g. filtertop10="1").

The client then just shows or hides elements based on if these data attributes
are set or not.

It's probably not *that* intensive for the client browser to do the sorting,
and it could potentially be more flexible (live-sorting newly added streams).
But it also sounds much more painful (for me personally) to integrate some
JS sorting library than to do it in Go. So that's why I made this choice /
tradeoff.

Todo: ensure that we grab streams from the DB in chronological order.