Dark mode
Click a link to switch the CSS between style.css and style_dark.css, which is a copy with inverted background and text colors.
I'll have to maintain this copy which will be a pain, but I don't know of another way to do it. Hooray CSS!
New versions of logo and search images.
Fixed search bug.