Clips WIP
Added Clips table in DB with stream id column. Along with insert and retrieve functions.
Added top clips endpoint to twitchapi.
In Update(), for each live channel, we also now query the clips of that last day and add any new ones to the DB.
Todo: CSS for clips, and choose only the top 2 viewed clips if there are >2.